:zap:  Refactor: intergrate transition into one type def and optimize type calculations at `@transition`
